% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment_statistics.R
\name{compute.ES}
\alias{compute.ES}
\title{compute the ES for a set}
\usage{
compute.ES(ranked.list, gene.set)
}
\arguments{
\item{ranked.list}{ranked list of objects}

\item{gene.set}{set of objects representing a process}
}
\value{
A list containing the ES of the set and it's position from the ES table
}
\description{
The enrichment score for a given gene set is calculated using a Kolmogorov-Smirnov (KS) test-like algorithm that walks down the ranked list of genes, increasing a running-sum statistic when a gene in the set is encountered and decreasing the statistic when a gene outside the set is encountered
}
\examples{
#ranked.list <- readRDS("data/ranked_list.rds")
#gene.set <- readRDS("data/one_gene_set.rds")
#compute.ES(ranked.list, gene.set)


}
