% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dotplot}
\alias{dotplot}
\title{Visualization functions for Enrichment analysis results
Function to plot a dot plot chart from the odd ration, the Fisher's test p.value and the normalized enrichment score of signficantly enriched èathways.}
\usage{
dotplot(data)
}
\arguments{
\item{data}{the resulting data frame from the multiple_set_analysis function}
}
\value{
a dotplot chart summarizing signficant results
}
\description{
Visualization functions for Enrichment analysis results
Function to plot a dot plot chart from the odd ration, the Fisher's test p.value and the normalized enrichment score of signficantly enriched èathways.
}
\examples{
ranked.list <- data(ranked.list)
pathways <- data(pathways)
data <- multiple_set_analysis(ranked.list, pathways)
dotplot(data)

}
