% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaf_file}
\alias{gaf_file}
\title{.gaf file containing GO for Homo sapiens}
\format{
GAF file
}
\source{
http://current.geneontology.org/products/pages/downloads.html
}
\usage{
gaf_file
}
\description{
A large dataset containing GO information for human in GMT format.
}
\examples{
data(gaf_file)

}
\keyword{datasets}
