% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmt_file}
\alias{gmt_file}
\title{.gmt file containing GO for Homo sapiens}
\format{
GMT file
}
\source{
https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp
}
\usage{
gmt_file
}
\description{
A large dataset containing GO information for human hallmark in GMT format.
}
\examples{
data(gmt_file)

}
\keyword{datasets}
