% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment_statistics.R
\name{max.deviation}
\alias{max.deviation}
\title{Maximum deviation from 0}
\usage{
\method{max}{deviation}(table)
}
\arguments{
\item{table}{A vector containing numerical values.}
}
\value{
The maximum deviation from 0 among the values in the table.
}
\description{
This function compute the maximum deviation from 0 of a vector of values.
}
\examples{
max.deviation(c(seq(1:10),seq(-13)))

}
