% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment_statistics.R
\name{multiple_set_analysis}
\alias{multiple_set_analysis}
\title{Run GSEA and Fisher test from leading-edge subset for multiple sets of object}
\usage{
multiple_set_analysis(ranked.list, pathways)
}
\arguments{
\item{ranked.list}{A ranked list of objects}

\item{pathways}{a data frame containing all sets names as gs_name in one column and the associated objects in another columns as gene_symbol}
}
\value{
A data frame with the following columns for each set of object :
   set = the set name.
   NES = the normalized ES for the set.
   p.value = The permutation test nominal pvalue for the ES.
   odd.ratio = The odd.ratio from the Fisher's test.
   fp.value = the pvalue of the Fisher's test.
}
\description{
Run GSEA and Fisher test from leading-edge subset for multiple sets of object
}
\examples{
#ranked.list <- readRDS("data/ranked_list.rds)
#pathways <- readRDS("data/pathways.rds")
#multiple_set_analysis(ranked.list, pathways)
}
