% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment_statistics.R
\name{one_set_analysis}
\alias{one_set_analysis}
\title{Permutation test. normalized scores and Fisher test from leading-edge subset for one set}
\usage{
one_set_analysis(ranked.list, gene.set, n.perm = 1000)
}
\arguments{
\item{ranked.list}{A ranked list of objects.}

\item{gene.set}{A set of objects representing a process.}

\item{n.perm}{The number of permutation test. By default fixed to 1000.}
}
\value{
A list containing:
   \item{ES.pvalue}{ES pvalue of the set.}
   \item{NES}{Normalized score for the set.}
   \item{odd.ratio}{Odd.ratio from the fisher test.}
   \item{fp.value}{P-value of fisher test.}
}
\description{
This function computes from the ES of a given set, some statistics to assess the significance of the set's ES.
}
\details{
To this purpose, a permutation test is performed to build an empirical normal distribution of ESs by shuffling the ranked list of genes. Then, the pvalue is computed by comparing
this distribution to the set ES. Also, the normalized ES for the set can be assess from the normal distribution similar to Z-score.
The second part of this function uses the leading-edge subset genes to run a Fisher's exact test. The leading-edge subset is fixed as the top genes participating to the ES of the set and are retrieved by
taking all the genes before reaching the maximum ES. The two factor compared are the presence of genes from the subset in the set and the presence of genes in the set not from the subset.
}
\examples{
#data(ranked.list)
#data(fatty_acid_metabolism)
#one_set_analysis(ranked.list,fatty_acid_metabolism)
}
