% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.oneset.gmt}
\alias{read.oneset.gmt}
\title{Read one set from .gmt file}
\usage{
read.oneset.gmt(set)
}
\arguments{
\item{set}{One line from .gmt file}
}
\value{
A data frame with the set name and the associated gene symbols
}
\description{
Format one set into a data frame : one set correspond to one line in .gmt format.
}
\examples{
lines <- readLines("data/ReactomePathways.gmt")
read.oneset.gmt(lines[1])

}
