% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_set_gmt}
\alias{read_set_gmt}
\title{Read whole .gmt file}
\usage{
read_set_gmt(path.file)
}
\arguments{
\item{path.file}{Path for the .gmt file}
}
\value{
A data frame containing sets with their corresponding gene symbols.
}
\description{
Read all lines through .gmt file and compute a data frame for each set to finally bind them in one total data frame.
}
\examples{
gmt <- data(gmt_file)
read_set_gmt(gmt)
}
