% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment_statistics.R
\name{compute.ES}
\alias{compute.ES}
\title{compute the ES for a set}
\usage{
compute.ES(ranked.list, gene.set)
}
\arguments{
\item{ranked.list}{ranked list of objects.}

\item{gene.set}{set of objects representing a process.}
}
\value{
A list containing the ES of the set and it's position in the ranked.list.
\item{ES}{The enrichment score for a given set.}
\item{ES.id}{The index of the maximum KS-cumulative sum for a given set (corresepond to the Enrichment score).}
}
\description{
The enrichment score for a given gene set is calculated using a Kolmogorov-Smirnov (KS) test-like algorithm.
}
\details{
This function fixes the weight for the running sum calculation as : sqrt((m.list - n.set)/n.set) for genes in the set and sqrt(n.set/(n.list - n.set)) in the other case.
Then, it goes through the ranked list of genes and sums the wieghts depedning if the gene is in or out the set.
Finally, it looks through the results of the cumulative sum and returns the maximum value deviating from 0 and it's index in the ranked list of genes.
}
\examples{
#data(ranked.list)
#data(fatty_acid_metabolism)
#compute.ES(ranked.list, fatty_acid_metabolism)


}
