% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_web_sets}
\alias{get_web_sets}
\title{Get sets of genes for Homo sapiens from Reactome, Gene ontology and KEGG from the web}
\usage{
get_web_sets()
}
\value{
A list containing different data frames with gene set symbols and associated genes symbols from the three databases.
   \item{[1] Reactome data frame}
   \item{[2] list of data frames containing different aspects for GO terms}
       \item{  [2][1] Cellulare component}
       \item{  [2][2] Biological process}
       \item{  [2][3] Molecular function}
   \item{[3] KEGG data frane}
}
\description{
This function uses different packages to access API of databases in order to retrieve gene sets.
Reactome is accessed from ReactomePA package.
GO terms are downloaded from the gene ontology database for different aspects.
KEGG is accessed with the KEGGREST package
}
\examples{
res <- get_web_sets()
str(res)
}
