% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment_statistics.R
\name{multiple_set_analysis}
\alias{multiple_set_analysis}
\title{Run GSEA and Fisher test from leading-edge subset for multiple sets of object}
\usage{
multiple_set_analysis(ranked.list, pathways, n.perm = 1000)
}
\arguments{
\item{ranked.list}{A ranked list of objects}

\item{pathways}{a data frame containing all sets names as gs_name in one column and the associated objects in another columns as gene_symbol}

\item{n.perm}{The number of permutation test. By default fixed to 1000.}
}
\value{
A data frame with the following columns for each set of object :
   \item{set}{The set name.}
   \item{NES}{The normalized ES for the set.}
   \item{p.value}{The permutation test nominal p-value for the ES.}
   \item{odd.ratio}{The odd.ratio from the Fisher's test.}
   \item{fp.value}{The p-value of the Fisher's test.}
}
\description{
This function is similar to one_set_analysis but generalized to many sets. It also correct the p-values for multiple testing.
}
\details{
The analysis run for each set is the same as the one_set_analysis function, but as many statistical tests are run for all the sets, this function
correct the p-values by using the build-in function p.adjust with the default method.
}
\examples{
#data(ranked.list)
#data(pathways)
#multiple_set_analysis(ranked.list, pathways)
}
